% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutoff.choose}
\alias{cutoff.choose}
\title{choosing cut}
\usage{
cutoff.choose(roc, col.nm, value)
}
\arguments{
\item{roc}{: num [1:99,1:6] roc table}

\item{col.nm}{: quantity of interest char}

\item{value}{: real number}
}
\description{
choosing cut
}
