% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Generating simulated data sets}
\usage{
data_generation(
  n_labeled,
  N_unlabeled,
  mean_1,
  sd_1,
  mean_2,
  sd_2,
  prevalence,
  setting = 1
)
}
\arguments{
\item{n_labeled}{the size of the labeled set}

\item{mean_1}{mean for controls}

\item{sd_1}{sd for controls}

\item{mean_2}{mean for cases}

\item{sd_2}{sd for cases}

\item{prevalence}{prevalence of cases}

\item{N_labeled}{the size of the unlabeled set}

\item{misspeci}{whether the model is correctly specified}
}
\value{
a simulated datasets containing outcome Y and phenotyping score for both labeled and unlabele sets
}
\description{
Generating simulated data sets
}
