% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolated_ROC}
\alias{interpolated_ROC}
\title{Interpolated ROC function used by supervised and ss imp estimators.}
\usage{
interpolated_ROC(S, Y, W = NULL, fpr_vals = seq(0.01, 0.99, by = 0.01))
}
\arguments{
\item{S}{Score}

\item{Y}{Outcome}

\item{W}{weight vector}

\item{fpr_vals}{equennce of fpr}
}
\value{
List containing:
\itemize{
  \item `cut`, sequence of cut
  \item `p.pos`, positive rate
  \item `fpr`, False positive rate
  \item `tpr`, True positive rate
  \item `ppv`, Positive predictive value
  \item `npv`, Negative predictive value
  \item `F.score`, F score test's accuracy
  \item `auc`, auc
}
}
\description{
Interpolated ROC function used by supervised and ss imp estimators.
}
