% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newfunction}
\alias{newfunction}
\title{choosing cut}
\usage{
newfunction(roc, col.nm, value)
}
\arguments{
\item{roc}{roc table}

\item{col.nm}{quantity of interest char}

\item{value}{desired value}
}
\description{
choosing cut
}
