% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outpt.FUN}
\alias{outpt.FUN}
\title{Helper functions for ROC estimation}
\usage{
outpt.FUN(tpr, fpr, p1, S, S.ini, fpr0 = seq(0.01, 0.99, by = 0.01))
}
\arguments{
\item{tpr}{True positive rate}

\item{fpr}{False positive rate}

\item{p1}{Actual postive rate}

\item{S}{Actual score}

\item{S.ini}{initial score}

\item{fpr0}{sequennce of fpr}
}
\value{
List containing:
\itemize{
  \item `auc`, auc
  \item `cut`, sequence of cut
  \item `p.pos`, positive rate
  \item `fpr`, False positive rate
  \item `tpr`, True positive rate
  \item `ppv`, Positive predictive value
  \item `npv`, Negative predictive value
  \item `F.score`, F score test's accuracy
}
}
\description{
Helper functions for ROC estimation
}
