% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertubation.R
\name{pertubation}
\alias{pertubation}
\title{Pertubation bootstrap for calculating the standard errors of the ROC estimates}
\usage{
pertubation(nbt, S_labeled, Y_labeled, S, Y, method)
}
\arguments{
\item{nbt}{number of replicates for pertubation boostrap}

\item{S_labeled}{phenotyping score in labeled set}

\item{Y_labeled}{outcome Y in labeled set}

\item{S}{phenotyping score in unlabeled set}

\item{Y}{outcome Y in unlabeled set}

\item{method}{validation methods, choosing from supervised and ssROC}
}
\value{
List containing AUC and ROC estimates for each perbutation bootstrap
}
\description{
Pertubation bootstrap for calculating the standard errors of the ROC estimates
}
