% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbation.R
\name{perturbation}
\alias{perturbation}
\title{Perturbation bootstrap for calculating the standard errors of the ROC estimates}
\usage{
perturbation(nbt, S, Y, method)
}
\arguments{
\item{nbt}{number of replicates for perturbation boostrap}

\item{S}{phenotyping score S for all, including labeled and unlabeled set}

\item{Y}{outcome Y for all, including labeled and unlabeled set; containing Y NA}

\item{method}{validation methods, choosing from supROC and ssROC}
}
\value{
list containing ROC estimates for each perturbation bootstrap
}
\description{
Perturbation bootstrap for calculating the standard errors of the ROC estimates
}
