% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_efficiency.R
\name{re}
\alias{re}
\title{Compute relative efficiency of supROC vs ssROC}
\usage{
re(sl_pert, ssl_pert, fpr_threshold)
}
\arguments{
\item{sl_pert}{perturbation result of supROC}

\item{ssl_pert}{pertubation result of ssROC}

\item{fpr_threshold}{FPR threshold to compute ROC parameters. Must choose 
from seq(0.01, 0.99, by = 0.01)}
}
\value{
vector of relative efficiency computed at each ROC parameter
}
\description{
Compute relative efficiency of supROC vs ssROC
}
