% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssROC.R
\name{ssROC}
\alias{ssROC}
\title{Imputation based semi-supervised method}
\usage{
ssROC(
  S,
  Y,
  W_labeled = NULL,
  W_unlabeled = NULL,
  fpr_vals = seq(0.01, 0.99, by = 0.01),
  bandwidth = NULL
)
}
\arguments{
\item{W_labeled}{optional vector of weights for labeled set}

\item{W_unlabeled}{optional vector of weights for unlabeled set}

\item{fpr_vals}{desired fpr sequence for output}

\item{bandwidth}{bandwidth for smoothing}

\item{S_all}{phenotyping score S for all, including labeled and unlabeled set}

\item{Y_all}{outcome Y for all, including labeled and unlabeled set; containing Y NA}
}
\value{
list containing
\itemize{
  \item `roc` roc
  \item `auc` auc
  \item `bandwidth` bandwidth for smoothing
  \item `mhat` estimate m
}
}
\description{
Imputation based semi-supervised method
}
