% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised.R
\name{supervised}
\alias{supervised}
\title{ROC analysis using supervised method}
\usage{
supervised(
  S_labeled,
  Y_labeled,
  W_labeled = NULL,
  fpr_vals = seq(0.01, 0.99, by = 0.01)
)
}
\arguments{
\item{S_labeled}{phenotyping score S in labeled set}

\item{Y_labeled}{outcome Y in labeled set}

\item{W_labeled}{optional vector of weights for labeled set}

\item{fpr_vals}{desired fpr sequence for output}
}
\value{
List containing AUC and ROC estimates
}
\description{
ROC analysis using supervised method
}
