% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Generating simulated data sets}
\usage{
data_generation(n_labeled, N_unlabeled, setting = 1)
}
\arguments{
\item{n_labeled}{the size of the labeled set}

\item{setting}{the simulation setting}

\item{N_labeled}{the size of the unlabeled set}
}
\value{
a simulated datasets containing outcome Y and S
}
\description{
Generating simulated data sets
}
